+++ pyinfra/modules/dnf.py
@@ -4,12 +4,10 @@ Manage dnf packages and repositories. Note that dnf package names are case-sensi
 
 from __future__ import unicode_literals
 
-from six import StringIO
-
 from pyinfra.api import operation
 
 from . import files
-from .util.packaging import ensure_packages, ensure_rpm
+from .util.packaging import ensure_packages, ensure_rpm, ensure_yum_repo
 
 
 @operation
@@ -39,23 +37,38 @@ def key(state, host, key):
 
 @operation
 def repo(
-    state, host, name, baseurl,
+    state, host, name, baseurl=None,
     present=True, description=None, enabled=True, gpgcheck=True, gpgkey=None,
 ):
+    # NOTE: if updating this docstring also update `yum.repo`
+    # COMPAT: on v1 rearrange baseurl/present kwargs
     '''
     Add/remove/update dnf repositories.
 
-    + name: filename for the repo (in ``/etc/dnf/repos.d/``)
-    + baseurl: the baseurl of the repo
+    + name: URL or name for the ``.repo``   file
+    + baseurl: the baseurl of the repo (if ``name`` is not a URL)
     + present: whether the ``.repo`` file should be present
     + description: optional verbose description
+    + enabled: whether this repo is enabled
     + gpgcheck: whether set ``gpgcheck=1``
     + gpgkey: the URL to the gpg key for this repo
 
-    Example:
+    ``Baseurl``/``description``/``gpgcheck``/``gpgkey``:
+        These are only valid when ``name`` is a filename (ie not a URL). This is
+        for manual construction of repository files. Use a URL to download and
+        install remote repository files.
+
+    Examples:
 
     .. code:: python
 
+        # Download a repository file
+        dnf.rpm(
+            {'Install Docker-CE repo via URL'},
+            'https://download.docker.com/linux/centos/docker-ce.repo',
+        )
+
+        # Create the repository file from baseurl/etc
         dnf.repo(
             {'Add the Docker CentOS repo'},
             name='DockerCE',
@@ -63,38 +76,16 @@ def repo(
         )
     '''
 
-    # Description defaults to name
-    description = description or name
-
-    filename = '/etc/yum.repos.d/{0}.repo'.format(name)
-
-    # If we don't want the repo, just remove any existing file
-    if not present:
-        yield files.file(state, host, filename, present=False)
-        return
-
-    # Build the repo file from string
-    repo_lines = [
-        '[{0}]'.format(name),
-        'name={0}'.format(description),
-        'baseurl={0}'.format(baseurl),
-        'enabled={0}'.format(1 if enabled else 0),
-        'gpgcheck={0}'.format(1 if gpgcheck else 0),
-    ]
-
-    if gpgkey:
-        repo_lines.append('gpgkey={0}'.format(gpgkey))
-
-    repo_lines.append('')
-    repo = '\n'.join(repo_lines)
-    repo = StringIO(repo)
-
-    # Ensure this is the file on the server
-    yield files.put(state, host, repo, filename)
+    yield ensure_yum_repo(
+        state, host, files,
+        name, baseurl, present, description, enabled, gpgcheck, gpgkey,
+        'dnf-config-manager',
+    )
 
 
 @operation
 def rpm(state, host, source, present=True):
+    # NOTE: if updating this docstring also update `yum.rpm`
     '''
     Add/remove ``.rpm`` file packages.