+++ pyinfra/modules/util/packaging.py
@@ -2,6 +2,7 @@ from __future__ import unicode_literals
 
 import six
 
+from six import StringIO
 from six.moves.urllib.parse import urlparse
 
 
@@ -160,3 +161,52 @@ def ensure_rpm(state, host, files, source, present, package_manager_command):
     # Package exists but we don't want?
     if exists and not present:
         yield '{0} remove -y {1}'.format(package_manager_command, info['name'])
+
+
+def ensure_yum_repo(
+    state, host, files,
+    name_or_url, baseurl, present, description, enabled, gpgcheck, gpgkey,
+    package_config_command,
+):
+    url = None
+    url_parts = urlparse(name_or_url)
+    if url_parts.scheme:
+        url = name_or_url
+        name_or_url = url_parts.path.split('/')[-1]
+        if name_or_url.endswith('.repo'):
+            name_or_url = name_or_url[:-5]
+
+    filename = '/etc/yum.repos.d/{0}.repo'.format(name_or_url)
+
+    # If we don't want the repo, just remove any existing file
+    if not present:
+        yield files.file(state, host, filename, present=False)
+        return
+
+    # If we're a URL, download the repo if it doesn't exist
+    if url:
+        if not host.fact.file(filename):
+            yield files.download(state, host, url, filename)
+        return
+
+    # Description defaults to name
+    description = description or name_or_url
+
+    # Build the repo file from string
+    repo_lines = [
+        '[{0}]'.format(name_or_url),
+        'name={0}'.format(description),
+        'baseurl={0}'.format(baseurl),
+        'enabled={0}'.format(1 if enabled else 0),
+        'gpgcheck={0}'.format(1 if gpgcheck else 0),
+    ]
+
+    if gpgkey:
+        repo_lines.append('gpgkey={0}'.format(gpgkey))
+
+    repo_lines.append('')
+    repo = '\n'.join(repo_lines)
+    repo = StringIO(repo)
+
+    # Ensure this is the file on the server
+    yield files.put(state, host, repo, filename)